Updated Chapter Summary.md
the updated version adds on at lines (65 to 70):
To add on, to ensure better immutability, an attribute/ variable should be declared as **final** such that the value cannot be modified after initialisation. 

For example, the mutation of this.x = 100 within a method of the same class will be prevented when the keyword **final** is used, adhering to the concept of immutability. Also, if the access modifier used is "protected" instead of "private", other classes within the same package will not be able to modify the attributes/variables declared as **final**.